refactor: split storage helpers out of persistence.js
persistence.js had grown to 1,409 lines and 70 methods covering path
resolution, file IO, rotation backups, snapshot export/import, migrations,
resets, diagnostics and the read/write core.

Three pieces come out of it:

- src/plugin/session-data.js holds the pure helpers. session-sync.js was
  carrying byte-identical copies of getPersistStamp() and hasSessionShape()
  (under the name isSessionDataShape), which put the rule deciding whether an
  incoming file is newer in two places - exactly the rule that must not drift
  when two devices disagree.

- storage-backup.js takes the hourly rotation backups, and
  storage-transfer.js the export/import snapshots. Both were self-contained
  and neither is on the path that a normal save takes.

The read/write core stays where it is. The five places where settings and
sessions could overwrite each other were just consolidated behind
writeSessionStore()/persistGlobalSettings(); spreading them back across files
would invite the same bug to grow back.

Also removes three dead methods: ensureStorageDir() (orphaned when vault-local
settings went away), the unused prototype wrapper around
normalizeSessionStorageLocation(), and isVersionHistoryCtrlRmbEnabled(), left
behind when the status bar gained configurable action slots.

persistence.js is down to 1,020 lines.
